Add db.js to handle db interaction
