+++ src/getProducts.js
@@ -14,6 +14,7 @@ import {
 } from "firebase/firestore";
 import $ from "jquery";
 import { hideLoading, showAlert, showLoading } from "./general";
+import { addProductToCurrentList } from "./db";
 
 async function displayProductsCards() {
   const productsRef = collection(db, "products");
@@ -147,48 +148,6 @@ function addProductData() {
     quantity: 1,
     unit: "pack",
   });
-}
-
-export async function addProductToCurrentList(product, productId) {
-  // get current user id
-  const userID = auth.currentUser?.uid;
-
-  // redirect to login if no user
-  if (!userID) {
-    window.location.href = "/login.html";
-  }
-  showLoading();
-  // get current list
-  try {
-    const queryRef = doc(db, "users", userID, "currentList", productId);
-    const querySnapshot = await getDoc(queryRef);
-
-    const productInCurrentList = querySnapshot.data();
-    if (!productInCurrentList) {
-      // add to current list
-      console.log("product is not in current list");
-      await setDoc(queryRef, {
-        productId: productId,
-        imageUrl: product.imageUrl,
-        name: product.name,
-        price: product.price,
-        count: 1,
-        sharedWith: [userID],
-      });
-    } else {
-      // update to current list
-      console.log("product is in current list");
-      await updateDoc(queryRef, {
-        count: productInCurrentList.count + 1,
-      });
-    }
-    showAlert("Product is added to your list", "warning");
-  } catch (error) {
-    showAlert("Something went wrong...", "warning");
-    console.log(error);
-  } finally {
-    hideLoading();
-  }
 }
 
 //Show previously added list
